+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - DEPRECATED: Old overlay component (uk-overlay-area-* uk-overlay-caption uk-overlay-toggle)
   - Added new classes to overlay component and reworked it completely
   - Added new hover classes to panel component
+  - Deactivated browser history navigation (for IE 11) in modal
 
   **Docs and Repo**