+++ CHANGELOG.md
@@ -3,10 +3,10 @@
 ### WIP
 
   - IMPORTANT: Removed caption component (Use overlay component)
-  - IMPORTANT: Removed uk-grid-preserve (Use uk-grid-large)
+  - IMPORTANT: Removed uk-grid-preserve
   - DEPRECATED: Old overlay component (uk-overlay-area-* uk-overlay-caption uk-overlay-toggle)
   - Reworked overlay component completely
-  - Added gutter collapse and large classes to grid component
+  - Added gutter collapse class to grid component
   - Added hover classes to panel component
   - Added new hover classes to panel component
   - Deactivated browser history navigation in modal for IE11